requirements.txt
accelerate==1.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
deepspeed==0.18.2
diffusers==0.35.2
filelock==3.20.0
fsspec==2025.10.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==25.0
pillow==12.0.0
psutil==7.1.3
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
sympy==1.13.1
tokenizers==0.22.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.57.1
triton==3.2.0
typing_extensions==4.15.0
urllib3==2.5.0
zipp==3.23.0
openai==2.8.1
peft==0.18.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.7